ノードエディタの保存と読み込み

+++ Src/Library/Json/nlohmann/json.hpp
@@ -18928,7 +18928,7 @@ class serializer
             }
         };
 
-        JSON_ASSERT(byte < utf8d.size());
+        JSON_ASSERT(size_t(byte) < utf8d.size());
         const std::uint8_t type = utf8d[byte];
 
         codep = (state != UTF8_ACCEPT)
